Refactor ClrMD for 3.0 (#1081)
* Rework ClrHeap to understand regions

Still a work in progress.

Added ClrGCHeap concept.

* Rename ClrGCHeap -> ClrSubHeap

* Rename MemorySearcher -> MemoryExtensions

* WIP - begin ClrObject refactor

* Clean up IClrTypeFactory/IClrObjectHelpers

* Finish cleaning up type helpers

* Completed ClrRuntime refactor

* Deleted more code

* Now compiling

* Test fixes

* Fix types

* Fix assert ordering

* Ensure heap flushes

* Fix heap issues

* Fix System.Object name

* More test fixes

* Fix generic names

* Fix dac tests

* Ensure final release of com objects on dispose

* Fix StackTrace.Thread

* Clean up handles and fields

* Clean up CrStackFrame

* Clean up types

* Fix type errors

* Remove failing test that never passed

* Remove roslyn analyzers

* Clean up suggestions

* Cleanups

* Code cleanup

* More code cleanup

* Fix nullable warnings

* More code cleanups

* Sealed/internal

* Implementation cleanup

* Update version

* Update WindowsThreadSuspender.cs

* Fix build errors

* Update DACNameParserTests.cs

* Add overlaps

+++ src/Microsoft.Diagnostics.Runtime/src/Extensions/PointerHelpers.cs
@@ -8,6 +8,6 @@ namespace Microsoft.Diagnostics.Runtime
 {
     internal static class PointerHelpers
     {
-        public static IntPtr AsIntPtr(this ulong address) => new IntPtr((nint)address);
+        public static IntPtr AsIntPtr(this ulong address) => new((nint)address);
     }
 }
