+++ src/Microsoft.Diagnostics.Runtime/src/Linux/ElfDynamicSection.cs
@@ -66,7 +66,7 @@ namespace Microsoft.Diagnostics.Runtime.Utilities
 
         public ulong SymbolTableVA { get; }
 
-        public ElfSymbolTable? SymbolTable {get; }
+        public ElfSymbolTable? SymbolTable { get; }
 
         public bool TryLookupSymbol(string symbolName, out ElfSymbol? symbol)
         {
@@ -77,9 +77,9 @@ namespace Microsoft.Diagnostics.Runtime.Utilities
             }
             try
             {
-                foreach (uint possibleLocation in GnuHash.GetPossibleSymbolIndex(symbolName))
+                foreach (int possibleLocation in GnuHash.GetPossibleSymbolIndex(symbolName))
                 {
-                    ElfSymbol s = SymbolTable.GetSymbol(possibleLocation);
+                    ElfSymbol s = SymbolTable.GetSymbol((uint)possibleLocation);
                     if (s.Name == symbolName)
                     {
                         symbol = s;
